+++ .changeset/short-lobsters-collect.md
@@ -1,0 +1,5 @@
+---
+'@modern-kit/utils': minor
+---
+
+feat(utils): pickBy, omitBy 함수 추가 - @ssi02014
